Separate PieceCount component
