+++ src/app/display/profile/profile.tsx
@@ -1,11 +1,11 @@
 import * as React from "react";
-import { ProgressBar } from "./progressBar";
+import { ProgressBar } from "../progressBar";
 import { MapStateToProps, connect, MapDispatchToProps } from "react-redux";
 import { AppState } from "@app/store";
 import { getXpToNextLevel } from "@common/utils";
-import { ownedPieceSelector } from "../store/pieceSelectors";
-import { buyXpAction } from "../store/actions/localPlayerActions";
+import { buyXpAction } from "../../store/actions/localPlayerActions";
 import { GamePhase, Constants } from "@common/models";
+import { PieceCount } from "./pieceCount";
 
 const renderProgressBar = (current: number, max: number) => `${current} / ${max} xp`;
 
@@ -13,7 +13,6 @@ interface ProfileStateProps {
     name: string;
     level: number;
     xp: number;
-    pieceCount: number;
     gameStarted: boolean;
 }
 
@@ -24,7 +23,7 @@ interface ProfileDispatchProps {
 type ProfileProps = ProfileStateProps & ProfileDispatchProps;
 
 const ProfileUnconnected: React.FunctionComponent<ProfileProps> = (props) => {
-    const { name, level, xp, pieceCount, gameStarted, onBuyXp } = props;
+    const { name, level, xp, gameStarted, onBuyXp } = props;
 
     if (gameStarted === false) {
         return null;
@@ -38,7 +37,9 @@ const ProfileUnconnected: React.FunctionComponent<ProfileProps> = (props) => {
 
             <p className="item level">Level {level}</p>
 
-            <p className="item pieces">{pieceCount} / {level} pieces</p>
+            <div className="item">
+                <PieceCount />
+            </div>
 
             <div className="level-bar">
                 <ProgressBar className="xp-progress" current={xp} max={xpForNextLevel} renderContents={renderProgressBar} />
@@ -52,8 +53,7 @@ const mapStateToProps: MapStateToProps<ProfileStateProps, {}, AppState> = state 
     name: state.localPlayer.name,
     level: state.localPlayer.level,
     xp: state.localPlayer.xp,
-    gameStarted: state.game.phase !== GamePhase.WAITING,
-    pieceCount: ownedPieceSelector(state).length
+    gameStarted: state.game.phase !== GamePhase.WAITING
 });
 
 const mapDispatchToProps: MapDispatchToProps<ProfileDispatchProps, {}> = dispatch => ({
